change from float32 float64

+++ binary/bench_test.go
@@ -11,7 +11,7 @@ import (
 func BenchmarkNative(b *testing.B) {
 	var s uint32 = 1000
 	bs := make([]byte, s*gltf.SizeOfElement(gltf.ComponentFloat, gltf.AccessorVec3))
-	data := make([][3]float32, s)
+	data := make([][3]float64, s)
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
 		for i := range data {
@@ -23,7 +23,7 @@ func BenchmarkNative(b *testing.B) {
 func BenchmarkWrite(b *testing.B) {
 	var s uint32 = 1000
 	bs := make([]byte, s*gltf.SizeOfElement(gltf.ComponentFloat, gltf.AccessorVec3))
-	data := make([][3]float32, s)
+	data := make([][3]float64, s)
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
 		Write(bs, 0, data)
@@ -33,7 +33,7 @@ func BenchmarkWrite(b *testing.B) {
 func BenchmarkWrite_builtint(b *testing.B) {
 	var s uint32 = 1000
 	bs := bytes.NewBuffer(make([]byte, s*gltf.SizeOfElement(gltf.ComponentFloat, gltf.AccessorVec3)))
-	data := make([][3]float32, s)
+	data := make([][3]float64, s)
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
 		binary.Write(bs, binary.LittleEndian, data)
